+++ src/js/constant.js
@@ -79,4 +79,5 @@ export const entity_to_icon = {
   recursive: "rotate-left",
   functionality: "screwdriver-wrench",
   preview: "eye",
+  filter_info: "magnifying-glass-plus",
 }